[SYCL] Handle nullptr arguments in level_zero
If specialization constant is not used on device, then the buffer for
it is not created on device side. In this case the host part passes nullptr
for the corresponding kernel operand holding the reference to the buffer
with constants.
This patch fixes the level_zero arguments handling to avoid SegFault
in the plugin.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>