[SYCL][CUDA] Add mem_advise reset and managed mem check (#5536)
This PR adds the mem_advise reset case and a managed memory check in cuda_piextUSMEnqueueMemAdvise.

cuMemAdvise only works on managed memory and returns an error if host or device memory is passed.
The SYCL-CTS tests mem_advise with host and device memory, which are not managed.  This PR prevents an error from being thrown in the cts usm test.

In addition, the sycl spec specifies for mem_advise `A value of 0 reverts the advice for ptr to the default behavior`, currently a value of 0 is treated as unknown and throws an error in the cts as well. This reset case is added to mem_advise.

This patch along with #5446 resolves issue #5209 

This change breaks the ABI