[Driver][SYCL] Do not pull in late static device libs for AMD GPU
The addition of the new OpenMP offloading driver path, which unbundles
archives later in the toolchain is interfering with the device linking
behaviors involving AMD GPU targets for SYCL offload compiles.

Disable the inclusion of these unbundled static archives as we do not yet
take advantage of the new offloading paths.