Restore getArgIndex in OCLTypeToSPIRV (#1567)
* Update for LLVM Optional API changes

Update for LLVM commit b5f8d42efe3e ("[ADT] Deprecate
Optional::{hasValue,getValue} (NFC)", 2022-08-07).

This is a mechanical replacement of `hasValue` to `has_value` and
`getValue` to `value`.

* Restore argument tracing in OCLTypeToSPIRV

Commit 4a9c78ee ("Prepare SPIRVWriter for type conversion without
opaque pointers. (#1499)", 2022-06-20) removed `getArgIndex` in favor
of passing `Idx`, but this leads to incorrect adaptation of argument
types if sampler arguments are not in the same position between
different functions.

We might be able to drop `adaptArgumentsBySamplerUse` entirely, as we
don't represent samplers as i32 anymore; but for now just fix the
regression.

Fixes https://github.com/KhronosGroup/SPIRV-LLVM-Translator/issues/1562

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/87f8a58