[SYCL] Use 0 for generic mem_advise (intel/llvm-test-suite#1140)
Some tests attempt to do a generic mem_advise to test certain behaviour
that should not be affected by the actual advice. Previously the CUDA
backend wouldn't correctly handle advice 0, which was supposed to
signify a reset of previously set advice. Now that the CUDA backend
correctly handles this option we can return to using 0. Likewise we can
PI enumerators from these tests and use 0 directly.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>