[SYCL] Improve sycl::event::get_profiling_info exception message if default constructed  (#6476)
The default constructed sycl::event would emit a misleading message when querying profiling information. The message would assume a queue is attached to such sycl::event, even if it's not the case. Added a condition to handle such case, and emit a more proper message, and updated appropriate unit test.