[SYCL][CUDA] Add concurrent memory check to mem_advice (#6567)
This PR adds a check to CUDA's mem advice to make sure the memory passed is managed memory. This change makes mem_advise line up with prefetch in terms of checking that both the device supports concurrent managed memory and that the memory passed to the function is managed.

This is to prevent cuMemAdvise from failing on the SYCL mem_advise CTS tests.