[CI] Update GPU driver uplift process
* Adding abilility to self test task on pull_request made
* Adding commit description containing driver version
* Removing linux_staging drivers update. linux_staging versions are used
to produce "unstable" containers in SYCL nightly workflow, but in
fact "unstable" matching "latest" currently and there are no testing
processes in the CI currently which use "unstable" containers. Not sure what
the original intention was, but until we have any *separate* testing processes
for "unstable" containers it doesn't make sense to update the versions.
* Use personal access token to kick off pre-commit testing automatically