[SYCL] E2E test for code_location.  (intel/llvm-test-suite#1123)
An e2e test for code_location is needed that tests on both host and kernel. Also, since it needs to be tested against NDEBUG that may disable asserts/XPTI/etc. so we just use output and FileCheck.

Note there is a possible bug in code_location where when we test from llvm-lit, the full path shows up for FileName. But when running directly, just the file name and extension is the value. Those tests are temporarily disabled here while this is being investigated.

Signed-off-by: Chris Perkins chris.perkins@intel.com