Drop the JointMatrixINTEL struct-renaming pass when opaque pointers are enabled. (#1570)
The frontend is being changed to lower the struct name to the correct LLVM name
directly, obviating the need for this check. See
https://github.com/intel/llvm/pull/6535 for this change.

This marks the removal of the final call to the deprecated method
Type::getPointerElementType, although there remains some code that is not fully
working with opaque pointers enabled.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/e2bb479