[SYCL][ABI-Break] Support accessor property interface
SYCL 2020 defines a common property interface for accessor and host_accessor.
Currently the DPC++ implementation of these classes take property list as a
constructor argument but does not store them. In order to add the remaining
interfaces the property lists must be stored on accessor and host_accessor,
which is likely to break ABI.