[SYCL] Emit a warning for floating-point size changes after implicit conversions (#6323)
Prior to this changes, we had the following 2 warnings for floating-point precision change:

```
-Wimplicit-float-conversion : (example) implicit conversion loses floating-point precision: 'double' to 'float'
-Wdouble-promotion          : (example) implicit conversion increases floating-point precision: 'float' to 'double'
```

Both the above warnings were turned off by default and only enabled when the corresponding warning flags were passed.
Also, both the warnings were only specific to Clang and were not enabled for SYCL (meaning device code).
If the floating point values remained the same on both the source and target semantics after the cast, no warning was emitted.

Example:
```
float PrecisionLoss = 1.1;  // implicit conversion loses floating-point precision: 'double' to 'float'
float SizeChange = 3.0;   // In this case, though we have assigned a double value to a 'float' variable, there is techncally no loss of precision and we don't get any warning here. 
But there is still a change in floating-point-rank (size) here ( from 'double' to 'float')
```
This PR identifies a floating-point size change after an implicit cast and triggers a warning when the -Wimplicit-float-size-conversion flag is passed. It is turned off by default.
This warning is enabled for both Clang and SYCL(i.e both host and device code)
The only other case this addresses is, if there is a precision loss and `-Wimplicit-float-size-conversion` is passed but `-Wimplicit-float-conversion` is not, we make sure to emit at least a size warning.