Allow passing a single MLIR file to cgeist. (#29)
Currently `cgeist` accepts one or more C/C++ files as inputs and lowers them to MLIR/LLVM_IR/.o based on the command line options passed to it.  This PR extends `cgeist` to allow passing a single MLIR file to it as input. This is useful for debugging because we might want to skip the AST->MLIR generation step and instead provide the MLIR to be lowered in a file so that we can easily edit it at will.