[SYCL] Fix cleanup of deferred commands
When we work in multiple threads and lock on graph failed because of its ownership by another thread - we add command to cleanup to the pool with deferred commands. Sometimes it happens that this pool is never cleaned and even in scheduler destructor it is ignored because of not fully correct check.
Current solution let scheduler to cleanup deferred commands in its destructor. Although in most cases it will be cleaned up much earlier because in any place where vector ToCleanup was empty - we will check if there is deferred commands awaiting for destruction.

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>