[SYCL] Move simple event_impl constructor to header (#6611)
On certain systems unittests that uses the event_impl constructor with
no arguments may disagree with the event_impl source file on the layout
of the default std::optional argument. This can cause unexpected memory
accesses when performing operations on the passed argument. This commit
works around this issue by moving the body of the constructor to the
header.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>