[SYCL] Add support for SYCL_EXTERNAL on device_global variables (#5963)
Device global spec declares that it is possible to apply `SYCL_EXTERNAL`
macro to `device_global` variables. This patch modifies `sycl_device`
attribute which is used to implement `SYCL_EXTERNAL` macro so it is
possible to apply `SYCL_EXTERNAL` macro to `device_global` variables and
it is not allowed to apply it to any other variables.