[SYCL] Enable discard_events mode for the Level Zero (#6533)
Teach the Level Zero plugin piEnqueue* functions to accept nullptr instead of a pointer to the output event. In this case event is created internally and is not visible externally.
Introduce RefCountExternal for pi_event which allows to track external references to an event. It allows to do some optimizations if we know that event is not externally visible. These optimizations are going to be implemented in subsequent PRs.
Don't create proxy event for batch if there are no externally visible events in the batch.