[SYCL] Properly apply a mask of supported queue properties in the OpenCL plugin (#6622)
"discard_events" property is passed to all backends during queue creation.
OpenCL runtime itself is not aware of such property so we need to apply
a mask of supported properties before calling OpenCL API for queue
creation. By mistake mask was applied only for OpenCL 2.0, this PR makes
it to be applied for OpenCL 1.2 as well.