[Driver][SYCL] Improve diagnostics when linking with mismatched objects (#6591)
When performing linking with -fsycl, it is required that the user provide
the -fsycl-targets=triple option so the driver knows which targets to
unbundle and use. There is no indication to the end user if they are using
a command line that matches up with the generated objects and archives.

Perform some additional checking against the archives and objects as they
are provided on the command line and cross check with what the user has
specified for targets. If these do not match, emit a diagnostic stating
as such and also list out the targets that were found in the objects.