[SYCL] Move accessor_impl to source directory
Moving impl part of host accessor implementation to avoid exposing
implementation details in the headers. This allows for more changes
in accessor without breaking ABI.

Also updated the gdb pretty-printer since it was relying on the
impl details which are not available for gdb(unless libsycl.so is
built with debug symbols) anymore. Instead of accessing members of
impl directly gdb printers now accessing helper methods. To prevent
compiler discarding these methods there are dummy references which
are active when NDEBUG is not defined.