[ESIMD] Fix saturation argument of DPAS
The template argument for saturation was declared/used such a way
that any type could be passed to it, which would cause enforcement of
saturation when not intended.
In even worse scenarios the DPAS call with 3 simd arguments was
recognized as DPAS with 2 simd arguments + saturation argument:
dpas(src0,src1,src2) was treated as dpas(src1,src2,sat),
which caused totally incorrect behavior at runtime.

Also, this patch fixes the incorrect detection of ops_per_channel for
tfloat32 type on HOST.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>