[SYCL] sycl::kernel::get_kernel_bundle() may return a kernel_bundle with a null impl.
This is an issue complaining about that return from getSyclObjImpl can be null.
I talked to @steffenlarsen off line and although this is considered to be impossible (every constructor should create for itself and impl), it doesn't seem to be enforced anywhere in the code. We decided that would be a good compromise to make sure that impl is always non-null.

Signed-off-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>