Create the DialectBuilder infrastructure (#41)
This PR introduces a `DialectBuilder` as a infrastructure to generate several MLIR dialects operations.
Currently we have a `MemRefBuilder` and a `LLVMBuilder` both derived from an abstract `DialectBuilder` class.
The 2 classes can be used to generate operations in the `MemRef` MLIR dialect and the `LLVM` MLIR dialect respectively.

Note: not all possible operations for either builder have been implemented in this PR.

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>