[SYCL] Add local_accessor and deprecate target::local (#6341)
This PR creates the `local_accessor` class by aliasing the accessor class with `target::local`. 
The motivation behind this is that `target::local` has been deprecated in favour of `local_accessor` in SYCL2020. 
The approach of aliasing is taken as the spec states that local_access has the same semantics and restrictions as accessor with target::local.

Related issue: #4713
llvm-test-suite: https://github.com/intel/llvm-test-suite/pull/1063