[Driver][SYCL] Add support for -fsycl-force-target
Introduce -fsycl-force-target=arg support.  This is used along with -fsycl
to allow the user to override the target used to unbundle device objects
from fat objects and archives.

For example, object.o is built from -fsycl-targets=spir64.  The user wants
to build with -fsycl-targets=spir64_gen and object.o.  Use of
-fsycl-force-target=spir64 allows for this to be accomplished.

Additional notes:
  - Only valid when used with a single triple from -fsycl-targets
  - Applies to all unbundled archives and objects