[SYCL][ABI-BREAK] Remove sycl::program class
According to the SYCL2020 spec the program class should be removed since it's
now deprecated and replaced by the new class kernel_bundle.
Removing of the program_impl class will be done with a separate commit
since it's not an abi-breaking change and some performance analysis should be
done in scope of that removal.