convert sycl.constructor(...) { Type = @id } (#45)
This PR implements the SYCLToLLVM conversion for lowering sycl::id<n>(....).
It introduces the class ConstructorPattern which generates an llvm.call to the appropriate ctor function of the sycl::id templated class when given a sycl.constructor(%1) {Type = @id} operation.

This PR also introduces a new too names sycl-mlir-opt which can be used to drive unit testing.

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>