Add complex group algorithms
Extends the implementation of group algorithms by adding support for
complex types. Operations that can be done element-wise are implemented
in headers and multiplication is implemented in CUDA PI. This involved
changing the macros in CUDA PI to accommodate complex and defining
complex types to use in PI interface. The scratch space for moving data
between workgroups in the PI also had to be doubled.