[SYCL][Reduction] Fix issue with multiple buffer reductions (#6680)
reduAuxCGFunc has different output target between the case when number
of work groups is one or more, while the kernel executed is the same. As
such we were re-using the same kernel code trying to adjust its name so
that a single one isn't used two times (#WGs check is a run-time one).

We used accessors' types for that but that approach only worked for USM
case (ptr vs buffer accessor for the output type depending on #WGs). In
case when the original reduction variable used a buffer, both type were
the same resulting in a "definition with same mangled name" error. Use
something that has distinct types to name the kernel.