[SYCL][CUDA] Enable CXX standard library funcs for CUDA backend (#6482)
This PR allows CXX stdlib funcs to be used for NVPTX backend. 

See https://github.com/intel/llvm/discussions/6379

llvm-test-suite test: https://github.com/intel/llvm-test-suite/pull/1112

It also adds the compiler flag `"-fbundle-no-offload-arch"`, which allows device code bundles to omit arch information, meaning a bundle with the tag `sycl-nvptx64-nvidia-cuda-sm_50` can be bundled with the tag `sycl-nvptx64-nvidia-cuda` when the flag is enabled. This allows for generic bc bundles for a given backend.

I have not moved the NVPTX complex implementation to libdevice. Let me know if I should do so.

The behaviour of `IsMathErrnoDefault()` now changes so that `-fmath-errno` is enabled by default for the CUDA toolchain. This allows libdevice funcs to be used by a similar exploit that the SYCL toolchain uses. If necessary this could be reset once the upstream discussion on NVPTX codegen for `llvm.sin` and other llvm builtins has been resolved and implemented. See https://discourse.llvm.org/t/nvptx-codegen-for-llvm-sin-and-friends/58170/11?u=jdoerfert