Support Intel FPGA simulator device selector
Create fpga_simulator_selector to be used to select the FPGA simulator.
This assumes that the user has compiled their program with -Xssimulation
(as well as -fintelfpga) to prepare for the simulator.

An object of class fpga_simulator_selector must be constructed early in
the execution of the host program in order to make the simulator
available at runtime.

The emulator and hardware FPGA device selectors will prefer the
emulator/hardware over a simulation device if both have the same device
name.