[SYCL] Improve sycl-post-link performance with -split=kernel (#6689)
Right now we are computing a new callgraph in every call to
extractCallGraph. extractCallGraph is called every time we do a module
split, so for -split=kernel, that would be once per kernel. For modules
with many kernels, this can take a very long time. We only need to
compute this once because the input IR doesn't seem to change between
splits.

This improves performance of sycl-post-link from ~45min to ~7min for an
example with 13k kernels

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>