[SYCL] Deprecate SYCL 1.2.1 device selectors (#6599)
Now that we have added the SYCL 2020 callable device selectors, we need
to prepare for the removal of the older SYCL 1.2.1 `device_selector`
class. The first step is to add the deprecation message to the 1.2.1
style device selectors, which this PR does. It also removes the usage of
those from our own codebase so as to not trip on our own messages in the
future.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>
Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Larsen, Steffen <steffen.larsen@intel.com>