[SYCL] Fix device comparison in removeDuplicateDevices
Kernel bundles attempt to remove duplicate devices from a passed device
list. This is done through the removeDuplicateDevices function which
creates a set through comparing devices by their native handles.
However, the getNative member function on device_impl used to get these
handles will retain the native devices if the backend is OpenCL. For
root devices this will not have an effect, but on sub-devices this can
potentially lead to a leak of the devices. As a fix this commit compares
the PI devices rather than the native handles.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>