[SYCL] Allow buitin_assume_aligned to be called from device code. (#6705)
Prior to this PR, if there is a direct call to __builtin_assume_aligned
from within the device code, we would get the following error:
`error: SYCL kernel cannot call a variadic function`

This PR allows `__builtin_assume_aligned` to be invoked from within
device code.
The reason for doing so is, for example, if we load a float* pointer
from a data structure, the compiler has no way of knowing the underlying
alignment of the data the pointer points to, and therefore it cannot
combine the load/stores and one ends up with many unnecessary load/store
instructions plus lots of unnecessary pointer arithmetic plus increased
register pressure. We see this all over the place in our kernels. With
`__builtin_assume_aligned` we can tell the compiler to safely assume a
certain alignment, therefore implicitly forcing coalescing.