[ESIMD] Do simd<T, N>* to <N x T> arg/ret type conv when possible.
Today, 'simd<T, N> foo(simd<T, N> x)' is codegenerated by clang as
  'void foo(simd<T, N>* sret(simd<T, N>) %res, simd<T, N>* %x)'
for the SPIRV target (unless __regcall is specified), which is then
converted to
  'void foo(<N x T>* sret(<N x T>) %res, <N x T>* %x)'
in the LowerESIMDVecArg, then to
  '<N x T> foo(<N x T> %x)'
in the VC BE.

With the opaque pointers this becomes impossible, and the optimization
must happen in the "ESIMD FE". This patch implements it.

Signed-off-by: Konstantin S Bobrovsky <konstantin.s.bobrovsky@intel.com>