[sycl-bisect] Add missing exit
This script has been missing one call to exit, which causes control to
fall through and run successful tests a second time when using --command
without --command-allow-bisect-codes. This didn't change the results of
the script, but it did make it slower if the test command takes a while
to run.