[SYCL-MLIR][SYCLToLLVM][DialectBuilder]: CallOp should not check whether func exits (#6772)
When generating a `func::CallOp` we currently try to lookup the func
declaration/definition in the MLIR module and assert if one is not
found. This is unnecessary because the function declaration/definition
might appear later in the module. Removing the assertion is fine
because, when the code is lowered to the LLVM dialect, if a call to a
inexistent function is found the program fails to compile cleanly.

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>