Add "if_device_architecture_is"
We decided not add extended aspects for the device architectures and to
add a separate enumeration instead.  This means that we will have both
`if_device_has` (which tests aspects) and `if_device_architecture_is`
(which tests the architecture).  The first phase of the implementation
plan will implement only `if_device_architecture_is` and only in AOT
mode.