[SYCL] Add --host-target flag and remove --arm flag (#6620)
This patch adds a new flag to `configure.py` to allow changing the host
target to build, this makes it easier to build on different hosts such
as ARM or PowerPC.

This patch also removes the `--arm` flag as it is now redundant,
`--host-target=ARM` or `--host-target=AArch64` should be used instead.

This is slightly different than the original `--arm` flag as it only
allows to enable one of the ARM platforms. But I'm not sure what the use
case was for enabling both, as only one will be the host platform for a
given build. And compilation time was given as a reason to drop the X86
architecture originally so only enabling the correct ARM architecture
should also help with that.

Co-authored-by: Alexey Bader <alexey.bader@intel.com>