[SYCL][Windows] Fix DataMovement test
Using `-g` with the regular `clang` command line is not supported on
Windows.

On Windows `clang-cl` and `/Mdd` should be used instead.

However it doesn't seem like this test is testing anything to do debug
info and I couldn't find any reason for having it in the history, so
removing `-g` is the simplest solution to make the test work on both
Linux and Windows.