[SYCL] Add assert for device_global without device_image_scope
device_global is currently not fully supported but for backends that
support it, they should be usable on device only when the
device_image_scope property is present. This commit adds a temporary
static assert to avoid prevent the use of device_global without
device_image_scope until proper initialization has been implemented.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>