[SYCL][FPGA] Support Intel FPGA simulator device selector (#6715)
Create **fpga_simulator_selector** to be used to select the FPGA
simulator device. This assumes that the user has compiled their program
with -Xssimulation (as well as -fintelfpga) to prepare for the
simulator.

An object of class fpga_simulator_selector must be constructed early in
the execution of the host program in order to make the simulator
available at runtime.

The current implementation has the effect that if an object of class
sycl::ext::intel::fpga_simulator_selector is defined, FPGA hardware
devices selected using sycl::ext::intel::fpga_selector will select
simulator devices. This will be documented in the release notes and the
specification. We expect this behaviour to be eliminated in the future.