[SYCL][CUDA] Add support of CUDA XPTI tracing (#6373)
Fork of https://github.com/intel/llvm/pull/5797

This patch rebases and finalizes (similarly to
https://github.com/intel/llvm/pull/6023) the draft in
https://github.com/intel/llvm/pull/5797, which already contained the
most important commits thanks to @alexbatashev.

The most relevant additions of this patch were done in CMake files, in
particular
 
- switch from `FindCUDA` CMake (deprecated) module to the
`FindCUDAToolkit` one in order to find cupti library by means of
`CUDA_cupti_LIBRARY`. This is advisable because on some systems
`FindCUDA` fails to find `CUDA_cupti_LIBRARY`. This is also the case of
the CI, see the
[log](https://github.com/intel/llvm/runs/7115612243?check_suite_focus=true)
in case of `FindCUDA` is used.
- find `generated_cuda_meta.h` for generating the CUDA printer
definitions, since the location of this header file seems to vary
depending on the system, in case of this file is not found a warning is
printed and no errors are thrown.

Co-authored-by: Alexander Batashev <alexbatashev@outlook.com>