[SYCL][PI/CL] check for device version and extensions
There might be platforms where the supported device version is lower
than the platform version or where not all devices do support the same
extensions and hence some extensions supported by particular devices
are not reported in the platform extensions.

In particular for CL_PLATFORM_EXTENSIONS the OpenCL specification says:
"[...] Each extension that is supported by all devices associated with
this platform must be reported here."

In "3.4.1 Mixed Version Support" the specification also says: "[...] The
version returned corresponds to the highest version of the OpenCL
specification for which the device is conformant, but is not higher than
the platform version."

Hence, check for the device version and extensions rather than the
platform version and extensions in piProgramCreate().

Signed-off-by: Danilo Krummrich <dakr@redhat.com>