[SYCLToLLVM]: Fix SubIndexOpLowering (#6797)
This PR fixes the lowering code for `polygeist.subindex`. Given:
```
 %0 = "polygeist.subindex"(%arg0, %c0) : (memref<?x!sycl_array_1_>, index) -> memref<1xi64>
```
the LLVM IR dialect code generated is (after conversion):
```
!sycl_array = !struct<"class.cl::sycl::detail::array.1", (array<1 x i64>)>
!ptr_sycl_array = !llvm.ptr<!sycl_array>
...
%0 = llvm.getelementptr %arg0[0, 0] : (!ptr_sycl_array) -> !llvm.ptr<i64>
```
The generated code is incorrect because the GEP yields an `array<1xi64>`
given that there are 2 indexes `[0,0]`.
The first zero indexes through the struct in `!sycl_array`, and the
second zero index yields the array. In order to produce a
`!llvm.ptr<i64>` (as requested by the polygeist subindex operation), the
generated GEP should be:
```
  %0 = llvm.getelementptr %arg0[0, 0, 0] : (!ptr_sycl_array) -> !llvm.ptr<i64>
```
Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>