[SYCL] Optimize operator[] method of host accessor
Compiler cannot optimize out calls to getPtr(), getOffset() and
getMemoryRange(), that were used in accessor::operator[], because
they may have side effects(e.g. printf). The patch adds a member
in the accessor which points to the actual data stored in the
accessor_impl. This pointer is initialized once during construction
and used instead of functions mentioned above.