[SYCL] Make sycl::vec trivially_copyable and align layout
Before the patch the data sycl::vec used underneath was either
cl_* types or clang vector types(aka ext_vector_type) depending
on the host compiler.

This was causing binary compatibility issues because such types
are passed differently according to the SysV calling convention.
The way such types are passed also depends on the instruction set
a compiler allowed to use(aka -avx).

The vector types were used underneath in order speed up sycl::vec
operations on the host when available(when compiling with clang).
So, to avoid performance degradation we need to continue doing the
operations using vector types.

The patch solves this problem by changing the underlying type to
std::array in all the cases for the host compilation. And adds
conversions to vector types when the host compiler is clang to
avoid loosing performance. The GCC vector_size could be used
to cover both compiler, but it has some problems to solve, e.g.
it does not support vector of lenght 3.

This also allows us to untie sycl::vec from OpenCL types.

The are things to improve:
1. Use vectors data type in more sycl::vec methods
2. Use vectors for half type
3. Refactor several ifdef SYCL_DEVICE_ONLY to if constexpr
   constructs.