[SYCL] Implement optional device feature diagnostics
This commit adds a warning diagnostic for when there is a mismatch
between aspect usage propagated to a function and the function's
`device_has` attribute. Additionally, notes accompany the warning to
give the user a trace through to where the aspects originate.
Since the aspects are propagated in an LLVM pass, the warning is issued
from there, which means no source information is available by default.
To alleviate this, Clang CodeGen will now add srcloc metadata with
an encoded version of the source location to allow the backend to
correctly report the location for the various new diagnostics.

Additionally, this commit also adds the SYCLPropagateAspectsUsage pass
to the passes run for SYCL device code.

Co-authored-by: Sabianin, Maksim <maksim.sabianin@intel.com>
Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>