[SYCL] Bump SYCL_MINOR_VERSION
Unlike previous releases we change the policy of bumping the version to
be done at the beginning of the development cycle instead of its end.
That allows customers who are using development snapshots to guard their
code in a way that would keep working after the actual release is done.