[SYCL][Doc] Specification & design for "if_device_has" (#6712)
Add specifications for two proposed extensions:

* "sycl_ext_oneapi_device_if": Allows device code to conditionally use
  "optional kernel features" based on the device's aspects.

* "sycl_ext_intel_device_architecture": Allows device code to
  conditionally use "optional kernel features" based on the device's
  architecture.

This PR also adds a design document describing how this can be
implemented.  However, the design is split into two phases, and the
document currently describes only the first phase.  We expect to update
this design soon(ish) to include the second phase.  There are several
limitations imposed by the first phase, and these are documented in the
extension specifications.