[Driver][SYCL] Add additional debug settings when using -g on Windows
A recent addition of adding sycld.lib on Windows when using the Linux based
driver was performed.  There is additional tie-ins needed for general usage
of debug libraries, which includes adding -D_DEBUG.
Add the needed macro and also pull in the debug runtime library to match.
This is done only for -fsycl based compilations as that is when we are
pulling in these particular libraries.