[SYCL-MLIR] Lower SYCL device module to LLVM
Introduce SYCL device module inside the host module, which should be
marked as a "gpu_container" if it contains device code. Also, the
device module is no longer held in an OwningOpRef, as it will be
removed by the lowering pass.

Define a temportal solution to lower GPU operations, which are lowered
as if they were their regular counterparts:

- GPUReturnOp: replace by an LLVM return (as in the original GPU to
LLVM pass);
- GPUModuleOp: Copy code to parent module and erase;
- ModuleEndOp: Erase;
- GPUFuncOp: Copy code to a new FuncOp in the host module and erase.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>