[SYCL] Fix debug info generation when integration footer is present (#6774)
This patch is to fix two known issues with debugging caused by
integration footer presence, without redesigning the integration footer
approach.
One issue is the missing checksum for the main file on Windows. This
causes the breakpoints not being hit.
The other issue being that the host compilation's DICompileUnit points
to the temporary
file created instead of the original source file. So file names and
checksums are different for host and compilation.

The changes made to fix the issues are:
1. Driver passes an absolute path to an original source file through
-main-file-name if integration footer is enabled.
2. Driver passes a full path to the original source file through
-full-main-file-name.
3. Clang CodeGen uses the file specified through -main-file-name and
full-main-file-name to calculate checksums and file names for the main
file.

Joint work with @AlexeySachkov

Signed-off-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>