[SYCL] Bump SYCL_MINOR_VERSION (#6855)
Unlike previous releases we change the policy of bumping the version to
be done at the beginning of the development cycle instead of its end.
That allows customers who are using development snapshots to guard their
code in a way that would keep working after the actual release is done.

The documentation for this is being done as
https://github.com/intel/llvm/pull/6856

We have consensus on how/when MAJOR/MINOR have to be bumped, so do it
now. There are some debates regarding PATCH/DEV versions updates so the
documentation would be updated later.