[SYCL][CUDA] Allow joint_matrix to be loaded from const T (#6532)
Fixes a bug where if `joint_matrix_load` attempts to load `joint_matrix`
from an array of `const T`incorrect behaviour will occur or an error
will be thrown. To fix this we make use of `std::remove_const_t<T>` in
appropriate places. This is important functionality for integrating
joint_matrix with existing SYCL-DNN routines.
I think that similar problems might occur in the intel backends for
their existing impl: I have not made corresponding changes because I do
not have the hardware to test it.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>