[sycl-post-link] Rework collection of global vars to extract from module in split mode
This patch helps to reduce size of the split modules and therefore
evident amount of debug info.
In case there are global variables which are not used in a split module, they
are now not included in such module. It also helps with the variables that are
not used in any function and are not furher removed with the Global DCE pass.