[SYCL] Change "static constexpr int" -> "constexpr int" (#6984)
in sycl/source/detail/image_impl.hpp. clang-cl and MSVC produce
different exports for this. However, the static here doesn't bring any
functional value, so just remove it as a quick solution to align the
behavior between two toolchains that could be used to compile the
project.

The change is expected to be NFC for the MSVC toolchain used in our CI.