[SYCL] Ignore some symbols during ABI check
MSVC and clang-cl have minor differences in how the exported symbols are
created under some circumstances. We hit some of them in SYCL RT but we
believe they don't affect customers and so could be excluded from the
ABI check so that it would pass when the library is compiled with either
toolchain.