[SYCL] Use -fsemantic-interposition for device_selector.cpp
We build the whole project with -fno-semantic-interposition by default
and it gets miscompiled by clang after https://github.com/intel/llvm/pull/6896.

It's not yet clear if this would be a permanent fix or if there is a bug
in clang. The issue is being tracked at https://github.com/llvm/llvm-project/issues/58295.