[SYCL] Relax checks in host_acc_opt LIT test
It turns out to be non-robust to expect compiler to vectorize the
code. The code can be optmized in different ways. So relax the
checks to verify that there are no calls to accessor helper methods
(getOffset and getMemoryRange) in the loop. Absence of these calls
should be enough to optimize the code.